requirements.txt
cprint                 
django-jazzmin         
Flask-Bcrypt           
Flask-Cors             
fuzzywuzzy             
jsonify                
mysql-connector-python 
